[CIR] Rename CIRBuildImpl->CIRGenModule, CIRBuildImpl->CIRGenModule, etc
CIRBuildImpl -> CIRGenModule
CIRContext -> CIRGenerator
CIRCodeGenFunction -> CIRGenFunction

This is to match clang's CodeGen a bit better. We'll need to refactor
some of the inline namespacing and private code in order to provide
access to members in the same exact fashion that codegen does and so we
might as well match their class patterns.