[CIR][CodeGen] Update loop codegen to use cir.brcond
Since 'cir.yield loopcondition' was removed, update to use cir.brconds.

This change was initially done because yielding a value out with loopcondition
require adding arguments to all regions involved, and the approach was somewhat
brittle. More importantly it was breaking branchop interfaces assumptions,
which was preventing the verifier to be used.